requirements.txt
transformers
torch
numpy
pandas
umap-learn
scikit-learn
matplotlib
